# packages/contracts/openapi/marib-tax.v1.yaml
openapi: 3.1.0
info:
  title: Marib Tax API
  version: 1.0.0
  description: Authoritative API-01 foundation. Business operations use the /api/v1 prefix.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:{port}
    description: Local development placeholder
    variables:
      port:
        default: '3000'
...
paths:
...
  /api/v1/masterdata/activities/{id}:
    parameters: [{ $ref: '#/components/parameters/ActivityId' }]
    get:
      operationId: getOwnedCommercialActivityById
      summary: Read an owned commercial activity summary
      tags: [Master data]
      x-permission: taxpayer.profile.read
      x-authorization-predicates: [OWNERSHIP]
      responses:
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '200':
          {
            description: Owned commercial activity,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        $ref: '#/components/schemas/CommercialActivitySummary',
                      },
                  },
              },
          }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/taxpayers/{id}:
    parameters: [{ $ref: '#/components/parameters/TaxpayerId' }]
    get:
      operationId: getOwnedTaxpayerById
      summary: Read an owned taxpayer profile summary
      tags: [Taxpayer registry]
      x-permission: taxpayer.profile.read
      x-authorization-predicates: [OWNERSHIP]
      responses:
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '200':
          {
            description: Owned taxpayer profile,
            content:
              {
                application/json:
                  {
                    schema:
                      { $ref: '#/components/schemas/TaxpayerProfileResponse' },
                  },
              },
          }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/visits:
    post:
      operationId: scheduleFieldVisit
      summary: Schedule a new field visit
      tags: [Field visits]
      x-permission: field_visit.schedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScheduleFieldVisitInput' }
      responses:
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '201':
          description: Visit scheduled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldVisitResponse' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/visits/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Uuid' }
    get:
      operationId: getFieldVisit
      summary: Get details of a field visit
      tags: [Field visits]
      x-permission: request.review
      responses:
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '200':
          description: Visit details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldVisitResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/visits/{id}/results:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Uuid' }
    post:
      operationId: recordFieldVisitResult
      summary: Record results of a field visit
      tags: [Field visits]
      x-permission: field_visit.result.record
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecordFieldVisitResultInput' }
      responses:
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '201':
          description: Result recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldVisitResultResponse' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/visits/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Uuid' }
    post:
      operationId: cancelFieldVisit
      summary: Cancel a scheduled field visit
      tags: [Field visits]
      x-permission: field_visit.schedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CancelFieldVisitInput' }
      responses:
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '200':
          description: Visit cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldVisitResponse' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Server-validated bearer token. No token or secret is stored here.
  schemas:
    Uuid:
      type: string
      format: uuid
...
    TaxpayerProfileResponse:
      type: object
      additionalProperties: false
      required:
        [
          id,
          publicRef,
          displayName,
          statusCode,
          hasTaxNumber,
          activeLegalEntityCount,
          openDuesFlag,
        ]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        publicRef: { type: [string, 'null'], minLength: 1 }
        displayName: { type: string, minLength: 1 }
        statusCode: { type: string, minLength: 1 }
        hasTaxNumber: { type: boolean }
        activeLegalEntityCount: { type: integer, minimum: 0 }
        openDuesFlag: { type: boolean }
...
    CommercialActivitySummary:
      type: object
      additionalProperties: false
      required: [id, publicRef, taxpayerId, name, statusCode]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        publicRef: { type: [string, 'null'], minLength: 1 }
        taxpayerId: { $ref: '#/components/schemas/Uuid' }
        name: { type: string, minLength: 1 }
        statusCode: { type: string, minLength: 1 }
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field: { type: string, minLength: 1 }
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
        message: { type: string, minLength: 1 }
    ErrorBody:
      type: object
      additionalProperties: false
      required: [code, message, traceId]
      properties:
        code: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
        message: { type: string, minLength: 1 }
        details:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
        traceId: { type: string, minLength: 1 }
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
...
    ScheduleFieldVisitInput:
      type: object
      additionalProperties: false
      required: [scheduledStartAt, scheduledEndAt, teamMemberStaffIds]
      properties:
        serviceRequestId: { type: [string, 'null'], format: uuid }
        balaghId: { type: [string, 'null'], format: uuid }
        scheduledStartAt: { type: string, format: date-time }
        scheduledEndAt: { type: string, format: date-time }
        teamMemberStaffIds:
          type: array
          items: { $ref: '#/components/schemas/Uuid' }
          minItems: 1
        locationSnapshot: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
    RecordFieldVisitResultInput:
      type: object
      additionalProperties: false
      required: [resultSummary, actualStartedAt, actualEndedAt]
      properties:
        resultSummary: { type: string, minLength: 1 }
        resultCode: { type: [string, 'null'] }
        actualStartedAt: { type: string, format: date-time }
        actualEndedAt: { type: string, format: date-time }
    CancelFieldVisitInput:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 1 }
    FieldVisitResponse:
      type: object
      additionalProperties: false
      required: [id, publicRef, serviceRequestId, balaghId, statusCode, actualStartedAt, actualEndedAt, locationSnapshot, notes, cancellationReason, createdByStaffProfileId, createdByProfileId, updatedAt, updatedByProfileId, archivedAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        publicRef: { type: [string, 'null'] }
        serviceRequestId: { type: [string, 'null'], format: uuid }
        balaghId: { type: [string, 'null'], format: uuid }
        statusCode: { type: string, enum: [scheduled, completed, cancelled] }
        actualStartedAt: { type: [string, 'null'], format: date-time }
        actualEndedAt: { type: [string, 'null'], format: date-time }
        locationSnapshot: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        cancellationReason: { type: [string, 'null'] }
        createdByStaffProfileId: { $ref: '#/components/schemas/Uuid' }
        createdByProfileId: { type: [string, 'null'], format: uuid }
        updatedAt: { type: [string, 'null'], format: date-time }
        updatedByProfileId: { type: [string, 'null'], format: uuid }
        archivedAt: { type: [string, 'null'], format: date-time }
    FieldVisitResultResponse:
      type: object
      additionalProperties: false
      required: [id, fieldVisitId, resultSummary, resultCode, recordedAt, recordedByStaffProfileId]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        fieldVisitId: { $ref: '#/components/schemas/Uuid' }
        resultSummary: { type: [string, 'null'] }
        resultCode: { type: [string, 'null'] }
        recordedAt: { type: string, format: date-time }
        recordedByStaffProfileId: { $ref: '#/components/schemas/Uuid' }
...
  parameters:
...
    TaxpayerId:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    ActivityId:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
  responses:
    Unauthenticated:
      description: A valid server-verifiable bearer identity is required.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: The verified actor is not authorized for this operation or resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ValidationFailed:
      description: The request body does not satisfy the published contract.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: The requested route or resource is unavailable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
security:
  - bearerAuth: []